backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
ultralytics==8.0.0
Pillow==10.0.0
